Corrected issue from VS update
Update changed the value of the MSVC Toolset Version, which affected the linker ability to find libraries such as <fstream> and <chrono>. Used the value of this field from the cuSolverSp example solution to resolve the issue.
Might need to abandon the idea of header file. Best to write functions in the main .cpp code? Would be nice to have a CPU-based check function to compare to GPU results.